+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### To Release
 
+- [editor:fix] c+d open document
 - [editor:fix] `space` will work for `text-field%` now
 - [editor:fix] `alt+backspace` will work for `text-field%` now